First Depth Test
Simple additions to the Kinect for Unity plugin that separates the depth
reading into specific ranges and sends a signal to Arduino depending on
the closest range, turning on a number of lights in relation to that
range.

Currently working on cleaning it up a bit more and fixing an issue with
the lights. Next up is determining where in the camera view the closest
depth is and how big it is (object detection)